requirements.txt
# Football Betting Agent - Requirements
# Python 3.10+

# Scraping
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0
playwright>=1.40.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
# tensorflow>=2.15.0  # optional, for neural network

# Statistics
scipy>=1.11.0
statsmodels>=0.14.0

# NLP (for news sentiment)
transformers>=4.35.0
nltk>=3.8.0

# Scheduling
apscheduler>=3.10.0

# Notifications
python-telegram-bot>=20.0

# Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
